元件上大致拆完;下一步開始寫一些跟後端有關的功能
