+++ next-client/pages/user/article-like.js
@@ -2,7 +2,7 @@ import Link from 'next/link'
 import React from 'react'
 import { Divider, Space, Tag } from 'antd'
 import UserSideBar from './user-side-bar'
-import UserSideBarMobile from './user-side-bar-mobile'
+import UserSideBarMobile from '../../components/user/user-side-bar-mobile'
 export default function Article() {
   let data = [
     {
