+++ next-client/pages/user/order/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 // import style from '@/styles/order.module.scss'
 import UserSideBar from '../user-side-bar'
-import UserSideBarMobile from '../user-side-bar-mobile'
+import UserSideBarMobile from '../../../components/user/user-side-bar-mobile'
 
 export default function Order() {
   return (
